+++ src/assets/ba_data/python/bauiv1lib/mainmenu.py
@@ -25,7 +25,7 @@ class MainMenuWindow(bui.MainWindow):
 
         # Preload some modules we use in a background thread so we won't
         # have a visual hitch when the user taps them.
-        bui.app.threadpool_submit_no_wait(self._preload_modules)
+        bui.app.threadpool.submit_no_wait(self._preload_modules)
 
         bui.set_analytics_screen('Main Menu')
         self._show_remote_app_info_on_first_launch()