+++ src/assets/ba_data/python/bauiv1lib/play.py
@@ -34,7 +34,7 @@ class PlayWindow(bui.MainWindow):
 
         # Preload some modules we use in a background thread so we won't
         # have a visual hitch when the user taps them.
-        bui.app.threadpool_submit_no_wait(self._preload_modules)
+        bui.app.threadpool.submit_no_wait(self._preload_modules)
 
         classic = bui.app.classic
         assert classic is not None