+++ src/assets/ba_data/python/bauiv1lib/settings/allsettings.py
@@ -26,7 +26,7 @@ class AllSettingsWindow(bui.MainWindow):
 
         # Preload some modules we use in a background thread so we won't
         # have a visual hitch when the user taps them.
-        bui.app.threadpool_submit_no_wait(self._preload_modules)
+        bui.app.threadpool.submit_no_wait(self._preload_modules)
 
         bui.set_analytics_screen('Settings Window')
         assert bui.app.classic is not None